refactor: modularize ScatterTimeSeriesChart by separating MainChart and OverviewChart components
- Introduced MainChart and OverviewChart components to enhance code organization and readability.
- Updated TimeSeriesChart to utilize the new components, improving maintainability and separation of concerns.
- Added downsampledData prop to TimeSeriesChart for better data handling.
- Refactored brush handling logic for improved clarity and functionality.
